+++ src/main/java/com/example/ecommerce/request/OrderItemRequest.java
@@ -3,15 +3,15 @@ package com.example.ecommerce.request;
 import java.math.BigDecimal;
 
 public class OrderItemRequest {
-    private Integer productId;
+    private Long productId;
     private Integer quantity;
     private BigDecimal price;
 
-    public Integer getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(Integer productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
